20.gün ödevi

+++ src/app/components/navi/navi.component.ts
@@ -1,5 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { FormBuilder,FormGroup,FormControl,Validators } from '@angular/forms';
+
 import { ToastrService } from 'ngx-toastr';
+
 import { Brand } from 'src/app/models/entities/brand';
 import { Color } from 'src/app/models/entities/color';
 import { BrandService } from 'src/app/services/brand.service';
@@ -14,19 +17,24 @@ export class NaviComponent implements OnInit {
 
   brands : Brand[] = [];
   colors : Color[] = [];
+  updateColorId:number = -1;
+  updateBrandId:number = -1;
   naviBrandFilterText : string = "";
   naviColorFilterText : string = "";
-  naviBrandNameText : string = "";
-  naviColorNameText : string = "";
+  brandForm : FormGroup;
+  colorForm : FormGroup;
   constructor(
     private brandService : BrandService,
     private colorService : ColorService, 
     private toastrService : ToastrService,
-    private dataSharingService : DataSharingService) { }
+    private dataSharingService : DataSharingService,
+    private formBuilder : FormBuilder) { }
 
   ngOnInit(): void {
     this.loadBrands();
     this.loadColors();
+    this.prepareBrandForm(null);
+    this.prepareColorForm(null);
   }
   loadBrands() {
     this.brandService.getBrands().subscribe(r=>{
@@ -50,18 +58,38 @@ export class NaviComponent implements OnInit {
       }
     });
   }
-  addBrand() {
-    let activeToast = this.toastrService.info("Lütfen Bekleyiniz","Bilgi");
-    let brand = {name:this.naviBrandNameText} as Brand;
-    this.brandService.addBrand(brand).subscribe(response=>{
-      if (response.success === true){
-        this.dataSharingService.isBrandAffected.next(true);
-        this.toastrService.remove(activeToast.toastId);
-        this.toastrService.success(response.message,"İşlem Başarılı");
+  insOrUpBrand(operationName:string) {
+    if (this.brandForm.valid){
+      let activeToast = this.toastrService.info("Lütfen Bekleyiniz","Bilgi");
+      let brandModel = Object.assign({},this.brandForm.value);
+      if (operationName === "Marka Güncelle"){
+        brandModel.id = this.updateBrandId;
+        this.brandService.updateBrand(brandModel).subscribe(response=>{
+          if (response.success === true){
+            this.dataSharingService.isBrandAffected.next(true);
+            this.toastrService.remove(activeToast.toastId);
+            this.toastrService.success(response.message,"İşlem Başarılı");
+          }else{
+            this.toastrService.error(response.message,"İşlem Başarısız");
+          }
+        });
+        this.updateBrandId = -1;
       }else{
-        this.toastrService.error(response.message,"İşlem Başarısız");
+        this.brandService.addBrand(brandModel).subscribe(response=>{
+          if (response.success === true){
+            this.dataSharingService.isBrandAffected.next(true);
+            this.toastrService.remove(activeToast.toastId);
+            this.toastrService.success(response.message,"İşlem Başarılı");
+          }else{
+            this.toastrService.error(response.message,"İşlem Başarısız");
+          }
+        });
       }
-    });
+      
+    }else{
+      this.toastrService.error("Kesinlikle Girilmesi Gereken Özellikler Girilmemiş.","Hata");
+    }
+    
   }
   deleteCurrentColor(color : Color) {
     let activeToast = this.toastrService.info("Lütfen Bekleyiniz","Bilgi");
@@ -76,17 +104,78 @@ export class NaviComponent implements OnInit {
       }
     });
   }
-  addColor() {
-    let activeToast = this.toastrService.info("Lütfen Bekleyiniz","Bilgi");
-    let color = {name:this.naviColorNameText} as Color;
-    this.colorService.addColor(color).subscribe(response=>{
-      this.dataSharingService.isColorAffected.next(true);
-      if (response.success === true){
-        this.toastrService.remove(activeToast.toastId);
-        this.toastrService.success(response.message,"İşlem Başarılı");
-      }else{
-        this.toastrService.error(response.message,"İşlem Başarısız");
+  insOrUpColor(operationName:string) {
+    if (this.colorForm.valid){
+      let activeToast = this.toastrService.info("Lütfen Bekleyiniz","Bilgi");
+      let color = Object.assign({},this.colorForm.value);
+      if (operationName === "Renk Güncelle"){
+        color.id = this.updateColorId;
+        this.colorService.updateColor(color).subscribe(response=>{
+          this.dataSharingService.isColorAffected.next(true);
+          if (response.success === true){
+            this.toastrService.remove(activeToast.toastId);
+            this.toastrService.success(response.message,"İşlem Başarılı");
+          }else{
+            this.toastrService.error(response.message,"İşlem Başarısız");
+          }
+        });
+        this.updateColorId = -1;
+        return;
       }
+      this.colorService.addColor(color).subscribe(response=>{
+        this.dataSharingService.isColorAffected.next(true);
+        if (response.success === true){
+          this.toastrService.remove(activeToast.toastId);
+          this.toastrService.success(response.message,"İşlem Başarılı");
+        }else{
+          this.toastrService.error(response.message,"İşlem Başarısız");
+        }
+      });
+    }else{
+      this.toastrService.error("Kesinlikle Girilmesi Gereken Özellikler Girilmemiş.","Hata");
+    }
+  }
+  prepareBrandForm(brand:Brand|null) {
+    if (brand === null){
+      this.brandForm = this.formBuilder.group({
+        name:["",Validators.required]
+      });
+      return;
+    }
+    this.brandForm = this.formBuilder.group({
+      name:[brand.name,Validators.required]
     });
+    this.updateBrandId=brand.id;
+  }
+  prepareColorForm(color:Color|null) {
+    console.log("Color: " + color);
+    if (color === null){
+      this.colorForm = this.formBuilder.group({
+        name:["",Validators.required]
+      });
+      return;
+    }
+    this.colorForm = this.formBuilder.group({
+      name:[color.name,Validators.required]
+    });
+    this.updateColorId=color.id;
+  }
+  getBrandModalTitle(){
+    if (this.updateColorId !== -1){
+      return "Marka Güncelle"
+    }
+    return "Marka Ekle"
+  }
+  getColorModalTitle(){
+    if (this.updateColorId !== -1){
+      return "Renk Güncelle"
+    }
+    return "Renk Ekle"
+  }
+  clearAllBrandCRUDCache() {
+    this.updateBrandId = -1;
+  }
+  clearAllColorCRUDCache() {
+    this.updateColorId = -1;
   }
 }
